Update use_case_description.md
The {} group paragraphs and are added to clarify the structure of the new text, shall be removed after review

The NEW: sections indicate the added text in order to incorporate the following feedback:

Good use case; well-written. You did not talk about the goal of the system though. It is a general time and task management calendar? Is it for instance targeted to students for them to balance their study load?